wip: cleaning up
