+++ lib/composer.js
@@ -501,14 +501,12 @@ class Composer {
     if (Array.isArray(merged)) {
       return merged.some((row) => {
         return followup.fields.some((f) => {
-          // console.log('is solved', f.schemaNode.name, Object.keys(row), (f.schemaNode.name in row))
           return (f.schemaNode.name in row)
         })
       })
     }
 
     return followup.fields.some((f) => {
-      // console.log('is solved', f.schemaNode.name, Object.keys(merged), (f.schemaNode.name in merged))
       return (f.schemaNode.name in merged)
     })
   }
@@ -519,7 +517,6 @@ class Composer {
  * @returns the merged partialResults
  */
 function mergeResults (query, partialResult, response) {
-
   let result = response[query.resolverName]
   let mergeAt = partialResult
   let mergeParent = null
@@ -565,7 +562,8 @@ function mergeResults (query, partialResult, response) {
       const r = result[i]
       // TODO! not by index!
       // match by source id
-      // TODO test with nullable values, rotate results
+      // TODO test with nullable values
+      // TODO test with rotated results
       for (const [k, v] of Object.entries(r)) {
         mergeAt[i][k] = v
       }
