+++ lib/query-builder.js
@@ -1,5 +1,6 @@
 'use strict'
 const { unwrapSchemaType, valueToArgumentString } = require('./graphql-utils')
+const { toQueryArguments, transitiveKeys, keySelection } = require('./utils')
 
 class QueryBuilder {
   constructor (options) {
@@ -193,7 +194,7 @@ class QueryBuilder {
                 info: this.info,
                 args,
                 types: this.types
-                // solved: false
+                // TODO solved: false
               }
               ctx.followups.set(node, followup)
             }
@@ -227,53 +228,12 @@ class QueryBuilder {
       }
     }
 
-    if (set.length < 1) {
-      console.warn('something wrong here, no fields in the query selection', keyFields)
-    }
+    // TODO
+    // if (set.length < 1) {
+    //   console.warn('something wrong here, no fields in the query selection', keyFields)
+    // }
 
     return `{ ${set.join(', ')} }`
-  }
-}
-
-// TODO move to utils
-function toQueryArguments (v, root = true) {
-  if (v === undefined) { return '' }
-
-  if (Array.isArray(v)) {
-    return `[${v.map(v => toQueryArguments(v, false)).join(', ')}]`
-  }
-
-  if (typeof v === 'object') {
-    const o = Object.keys(v).map(k => `${k}: ${toQueryArguments(v[k], false)}`)
-    return root ? o : `{ ${o} }`
-  }
-
-  // TODO test: quotes
-  return typeof v === 'string' ? `"${v}"` : v.toString()
-}
-
-// TODO move to utils
-// TODO performance
-// function keySelection(path) {
-//   if (path.indexOf('.') === -1) { return path }
-
-//   const nested = path.split('.')
-//   const root = nested.shift()
-//   return nested.reduce((result, k) => `${result} {${k}}`, root)
-// }
-
-function keySelection (path) {
-  if (path.indexOf('.') === -1) { return path }
-
-  return path.split('.').pop()
-}
-
-function transitiveKeys (type, subgraphEntities) {
-  for (const entity of Object.values(subgraphEntities)) {
-    const key = entity.keys.find(k => k.type === type)
-    if (key) {
-      return [key]
-    }
   }
 }
